Now that upstream has done quite a bit of work to get facter to work with ruby 1.9 we need to make the Debian package work with ruby1.9.
This came about because there was a lintian problem:

E: facter: ruby-script-but-no-ruby-dep ./usr/bin/facter

That happened because the install.rb, provided by facter, munges the
perfectly fine shebang in bin/facter (#!/usr/bin/env ruby) into the
ruby1.8 specific one (#!/usr/bin/ruby1.8) just because my
/usr/bin/ruby is a symlink to /usr/bin/ruby1.8.

Because facter can work with both 1.8 and 1.9, having a shebang that
specifically uses 1.8, when someone might have only 1.9 installed is
actually an error.

So, after some investigation, I found that the way to do this is to
switch to the newer setup.rb mechanism for installation (as install.rb
has been deprecated by Minero Aoki in favor of setup.rb for years)
which allows for a switch that keeps the shebang from being
re-written.

So, to summarize, the changes I made to get this to work:

 * Added an optional dependency on the libopenssl-ruby1.9 package
 * Stopped using the upstream provided install.rb (which btw. has been
   deprecated by Minero Aoki years ago) and switched to using the
   standardized Debian ruby CDBS classes that provide a setup.rb
 * Shipping the facter libraries into /usr/lib/ruby/vendor_ruby instead
   of /usr/lib/ruby/ruby1.8 (NB: the vendor_ruby directory seems to be
   the one the Debian Ruby team intends to use for version independent
   ruby libraries, as most recently discussed here:
   http://www.mail-archive.com/debian-ruby@lists.debian.org/msg00528.html)