fixes #2573, #2085, #1291 - fixes domain and fqdn facts resolution
This patch removes the relationship between the domain fact and LDAP/NIS domains.
domain fact relates to DNS domain - this will avoid the confusion caused
by the LDAP/NIS domain (which might be different to the DNS domain name).
Additionally, if hostname is already in long form, it won't try to build
the fqdn fact from hostname and domain.